+++ Examples/SciChartDemo/SciChartDemo/Examples/Examples3D/RealTime3DCharts/RealtimeUniformMesh3DChartView.m
@@ -89,7 +89,7 @@
                 double d = M_PI * radius * freq;
                 double value = sin(d) / d;
                 
-                int index = [indexCalculator getIndexAtUIndex:i andVIndex:j];
+                NSInteger index = [indexCalculator getIndexAtUIndex:i andVIndex:j];
                 items[index] = isnan(value) ? 1 : value;
             }
         }
